[#31] Convert Catalog into class instead
The use of Catalog as a database collection was getting too tedious. The requisite string nature collected from the OpenAPI made the regex parsing a total chore and even when everything seemed okay there were always little inconsistencies or inaccuracies. Instead of trying to keep everything as a database collection,  I opted to store all the information inside a class file that way we can keep track of the various tweaks we must continuously introduce as opposed to invisibly mutate the database collection and not see where a potential break could occur.