[#31] Finish basic implementation of findCandidateTermSchedules
Trying to find all possible combinations of non-specified courses (i.e. of all candidates courses) might be too many possibilities and might not only make the options too numerous to display but severely hamper the speed. To this end, the best/easiest way to proceed was to force the user to select all courses they wish to register for each term and us not try to propose a million options.

For example, if the user specifies 4 courses to take in a term, and say each for each course there's 24, 2, 36, and 36 combinations of sections between them (different arrangements for LEC+TUT+LAB for the course), that makes for a total of ~24*2*36*36 = 62,208 possible combinations of schedules for that given term (mind you many will likely have time conflicts). Now imagine if we were to "suggest" courses to take if the user didn't specify enough courses to take, that would involve so many possibilities that it would be unrealistic to compute in a short enough time and to even be meaningful to the user when trying to observe them.